Add the HideIcon attribute to automatically disable scene view icons
Unity automatically displays icons in the scene view for anything that has a custom icon set.

Nominally, you would just use GizmoUtility.SetIconEnabled to turn off icons. However,
this only works if Unity has created annotation data for the relevant type. It does
this when the annotation window (the Gizmos drop-down) is opened.

To mimic this in my own code, I had to use reflection to access AnnotationUtility.

The Core package now scans for types that have the HideIcon attribute but that
aren't being kept track of yet, refreshes annotation data, and then turns off
the relevant icons. This way, it *only* turns off icons once -- it will not
interfere if you turn them back on manually.

This only takes a few milliseconds to do, so it seems acceptable. I may
add a way to turn this off later, though, since it's still a bit invasive feeling.